[*] checkpoint

+++ ctrl_sw/src/power.h
@@ -8,7 +8,7 @@
 
 inline void pwr_sleep(void) {
                     rfm12b_sleep();
-                    CHRG_EN_GPIO_Port->ODR &= ~CHRG_EN_Pin;
+                    CHRG_EN_GPIO_Port->ODR &= (uint32_t)~CHRG_EN_Pin;
                     HAL_PWR_DisableWakeUpPin( PWR_WAKEUP_PIN1 );
                     __HAL_PWR_CLEAR_FLAG(PWR_FLAG_WU);
                     HAL_PWR_EnableWakeUpPin(PWR_WAKEUP_PIN1);
